Added all UI's

+++ src/org/wolf/crud/TestCRUD.java
@@ -29,11 +29,11 @@ public final class TestCRUD {
 		return null;
 	}
 	
-	public static Integer insertTest(String test_name, String lab_name, Integer specialized_doctor, Integer fees){
+	public static Integer insertTest(String test_name, String lab_name, Integer specialized_doctor, Double fees){
 		return null;
 	}
 	
-	public static Boolean updateTest(Integer test_id, String test_name, String lab_name, Integer specialized_doctor, Integer fees){
+	public static Boolean updateTest(Integer test_id, String test_name, String lab_name, Integer specialized_doctor, Double fees){
 		return null;
 	}
